+++ app/page.js
@@ -1,5 +1,0 @@
-import App from '../property-management-app.jsx'
-
-export default function Page() {
-  return <App />
-}